requirements.txt
fastapi==0.89.1
uvicorn==0.20.0
sqlalchemy==1.4.46
python-dotenv==0.21.1
psycopg2-binary==2.9.5
